+++ scheduler/tests/test_task_types/test_once_task.py
@@ -1,5 +1,6 @@
-from datetime import timedelta
+from datetime import timedelta, datetime
 
+from django.core.exceptions import ValidationError
 from django.utils import timezone
 
 from scheduler import settings
@@ -17,6 +18,21 @@ class TestScheduledTask(BaseTestCases.TestSchedulableTask):
         job.queue = self.queue_name
         job.callable = "scheduler.tests.jobs.test_job"
         self.assertIsNone(job.clean())
+
+    def test_create_without_date__fail(self):
+        task = task_factory(self.task_type, scheduled_time=None, instance_only=True)
+        self.assertIsNone(task.scheduled_time)
+        with self.assertRaises(Exception) as cm:
+            task.clean()
+        self.assertTrue(isinstance(cm.exception, ValidationError))
+        self.assertEqual(str(cm.exception), "{'scheduled_time': ['Scheduled time is required']}")
+
+    def test_create_with_date_in_the_past__fail(self):
+        task = task_factory(self.task_type, scheduled_time=datetime.now() - timedelta(days=1), instance_only=True)
+        with self.assertRaises(Exception) as cm:
+            task.clean()
+        self.assertTrue(isinstance(cm.exception, ValidationError))
+        self.assertEqual(str(cm.exception), "{'scheduled_time': ['Scheduled time must be in the future']}")
 
     def test_unschedulable_old_job(self):
         job = task_factory(self.task_type, scheduled_time=timezone.now() - timedelta(hours=1))